refactor: simplify platform support in require package
